Update Documentation (#379)
* Update Docs

* Add additional comments to be picked up by docfx

* Fix docfx to see new lib location.

* Clean up

* Update index.md

* Update docfx.json

* Update CertificateObject.cs

* Update CommandOptions.cs

* Update CollectObject.cs

* EventLogObjects now use a proper timestamp

* Better document ComObject

* use dllcharacteristics natively instead of tostring

* Document EventLogObject

* Collect AddressFamily and Transport Enums instead of strings.

* More comments

* Separate some classes into their own files.

* Update Settings.cs

* Modify how RowKeys work so objects could implement unique row keys per object if needed.

* Performance improvements by shortcutting operators when appropriate.

* Document + Clarity refactor of Analyzer.cs

* Parsing bug in analyzer.cs

* Update Analyzer.cs

* Missing ;s

* Build fixes.

* Run GenerateDocfx

* Fixes

* Update README.md

* Cache clause evaluations

* Fix off by 1

* Use our own Enum

* Update CollectObject.cs

* Update AnalyzerExpressionsTests.cs

* Add Address Family back

* Switch to newtonsoft.json

* Update docs

* Fix Serialization

* Fix LGTM Issues

* More LGTM Fixes

* Use an actual random password for tests

The pipeline systems require a certian password complexity that just a GUID wasn't matching.

* Improve RunExternalCommand

* Reduce maximum length to avoid interactive prompt

* Update CollectorTests.cs

* Update pr-validation.yml

+++ Lib/Utils/CryptoHelpers.cs
@@ -53,7 +53,7 @@ namespace AttackSurfaceAnalyzer.Utils
         {
             var bytes = new byte[8];
             crypto.GetBytes(bytes);
-            return (BitConverter.ToUInt64(bytes, 0) >> 11) / ulong.MaxValue * max;
+            return (BitConverter.ToUInt64(bytes, 0) >> 11) / (double)ulong.MaxValue * max;
         }
 
     }
